refactor: Legacy MVC 제거 및 DispatcherServlet 패키지 이동
