+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerAdapterRegistryTest.java
@@ -4,38 +4,16 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.tobe.handlerAdapter.AnnotationHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.handlerAdapter.HandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.handlerAdapter.HandlerAdapterRegistry;
-import com.interface21.webmvc.servlet.mvc.tobe.handlerAdapter.ManualHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.handlerMapping.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.mvc.tobe.handlerMapping.HandlerExecution;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class HandlerAdapterRegistryTest {
-
-    class LegacyController implements Controller {
-        @Override
-        public String execute(HttpServletRequest req, HttpServletResponse res) {
-            return null;
-        }
-    }
-
-    @DisplayName("레거시 MVC 컨트롤러를 실행할 수 있는 핸들러 어댑터를 조회한다.")
-    @Test
-    void getManualHandlerAdapter() {
-        HandlerAdapterRegistry handlerAdapterRegistry = new HandlerAdapterRegistry();
-        handlerAdapterRegistry.addHandlerAdapter(new ManualHandlerAdapter());
-        handlerAdapterRegistry.addHandlerAdapter(new AnnotationHandlerAdapter());
-
-        HandlerAdapter handlerAdapter = handlerAdapterRegistry.getHandlerAdapter(new LegacyController());
-
-        assertThat(handlerAdapter).isExactlyInstanceOf(ManualHandlerAdapter.class);
-    }
 
     @DisplayName("어노테이션 기반 컨트롤러를 실행할 수 있는 핸들러 어댑터를 조회한다.")
     @Test
@@ -49,7 +27,6 @@ class HandlerAdapterRegistryTest {
         when(request.getMethod()).thenReturn("GET");
 
         HandlerAdapterRegistry handlerAdapterRegistry = new HandlerAdapterRegistry();
-        handlerAdapterRegistry.addHandlerAdapter(new ManualHandlerAdapter());
         handlerAdapterRegistry.addHandlerAdapter(new AnnotationHandlerAdapter());
 
         HandlerExecution handler = annotationHandlerMapping.getHandler(request);
